Extend tests

+++ Tests/IntegrationTests/Satchels_Test.cs
@@ -29,29 +29,44 @@ namespace Tests.IntegrationTests
             //Assert export succesfull
             Assert.IsTrue(File.Exists(filePath));
 
+            CodexCollection? deserializedCollection = null;
             CodexCollection? importedCollection = null;
             try
             {
                 //Wipe to collection from COMPASS
                 MainViewModel.CollectionVM.DeleteCollection(testCollection);
 
-                //Import
-                importedCollection = await IOService.OpenSatchel(filePath);
-                Assert.IsNotNull(importedCollection);
-                ImportCollectionViewModel importViewModel = new(importedCollection);
+                //Deserialize Satchel
+                deserializedCollection = await IOService.OpenSatchel(filePath);
+                Assert.IsNotNull(deserializedCollection);
+                ImportCollectionViewModel importViewModel = new(deserializedCollection);
 
-                Assert.IsTrue(importViewModel.ContentSelectorVM.HasCodices, "imported Collection has no Codices");
-                Assert.IsTrue(importViewModel.ContentSelectorVM.HasTags, "imported Collection has no Tags");
-                Assert.AreEqual(testCollection.AllCodices.Count, importedCollection.AllCodices.Count);
-                Assert.AreEqual(testCollection.AllTags.Count, importedCollection.AllTags.Count);
+                Assert.IsTrue(importViewModel.ContentSelectorVM.HasCodices, "deserialized satchel has no Codices");
+                Assert.IsTrue(importViewModel.ContentSelectorVM.HasTags, "deserialized satchel has no Tags");
+                Assert.AreEqual(testCollection.AllCodices.Count, deserializedCollection.AllCodices.Count);
+                Assert.AreEqual(testCollection.AllTags.Count, deserializedCollection.AllTags.Count);
+
+                //Complete import to new Collection
+                importViewModel.CollectionName = "Imported_Satchel"; //cannot use a protect __ name because it is an illegal name
+                importViewModel.MergeIntoCollection = false;
+                await importViewModel.Finish();
+
+                importedCollection = MainViewModel.CollectionVM.CurrentCollection;
+
+                Assert.AreEqual(deserializedCollection.AllCodices.Count, importedCollection.AllCodices.Count);
+                Assert.AreEqual(deserializedCollection.AllTags.Count, importedCollection.AllTags.Count);
             }
             finally
             {
                 //Cleanup
                 MainViewModel.CollectionVM.DeleteCollection(testCollection);
-                if (importedCollection != null)
+                if (importedCollection != null && importedCollection.DirectoryName == "Imported_Satchel") //make sure not to delete something else
                 {
                     MainViewModel.CollectionVM.DeleteCollection(importedCollection);
+                }
+                if (deserializedCollection != null && deserializedCollection.DirectoryName.StartsWith("__")) //make sure not to delete something else
+                {
+                    MainViewModel.CollectionVM.DeleteCollection(deserializedCollection);
                 }
                 File.Delete(filePath);
             }
